+++ src/hfm/hfm.h
@@ -19,7 +19,7 @@ typedef struct HfmCode_s {
 
 typedef unsigned char HfmWord;
 
-extern uint32_t hfm_compress_block(CodeTable table, BlockHeader *header, uint8_t *source, uint8_t *dest, uint32_t n);
-extern uint32_t hfm_decompress_block(CodeTable table, BlockHeader *header, uint8_t *source, uint8_t *dest);
+extern uint16_t hfm_compress_block(CodeTable table, BlockHeader *header, uint8_t *source, uint8_t *dest, uint16_t n);
+extern uint16_t hfm_decompress_block(CodeTable table, BlockHeader *header, uint8_t *source, uint8_t *dest);
 
 #endif // HFM_H
